refactor(client): rename sheet components to kebab-case
Renames three sheet SFCs to match the kebab-case filename convention
used by sibling components in the same directories, and updates every
import path that referenced them. No component internals changed.

- common/Sheet.vue -> common/sheet.vue
- calendar-management/FundingSheet.vue -> calendar-management/funding-sheet.vue
- calendar/CreateCalendarSheet.vue -> calendar/create-calendar-sheet.vue
