+++ src/client/components/logged_in/calendar/create-calendar-sheet.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { useTranslation } from 'i18next-vue';
-import Sheet from '@/client/components/common/Sheet.vue';
+import Sheet from '@/client/components/common/sheet.vue';
 import CreateCalendarForm from './CreateCalendarForm.vue';
 
 const emit = defineEmits<{
